TONS of stuff... I really should have committed incrementally. Anyway...
New packages for jcs, and its dependency isp
New package for easymock
Mx4j now blocks a certain version of resin. This is really temporary, because we should be using hessian/burlap instead of resin (they're from the same devs). This is the only way to get it emerged for the moment

Re-slotted a ton of packages. So, you may need to re-emerge things if you start
seeing error messages from jar-from


svn path=/gentoo-java-experimental/; revision=548