sys-apps/baselayout-java: Apply patches from main tree to 9999 source ebuild
The ebuild for baselayout-java in the main tree has received some patches,
including the migration from the no longer available fdo-mime eclass to
the xdg-utils eclass. Moreover, the source repository for baselayout-java
was migrated from SVN to GIT.

This commit, hence, switches the source ebuild in the Java overlay from
the SVN to the GIT repository, and applies the patches from the ebuild
in the main tree to the source ebuild in the Java overlay.

Closes: https://github.com/gentoo/java-overlay/issues/48
Closes: https://bugs.gentoo.org/760474
Signed-off-by: Henning Sudbrock <github@hsudbrock.de>